Add version and header writing
