+++ inkcpp/story_impl.h
@@ -4,6 +4,7 @@
 #include <config.h>
 #include "types.h"
 #include "story.h"
+#include "header.h"
 
 namespace ink::runtime::internal
 {
@@ -11,26 +12,6 @@ namespace ink::runtime::internal
 	class story_impl : public story
 	{
 	public:
-		struct Header {
-			static Header parse_header(const char* data);
-
-			template<typename T>
-			static T swap_bytes(const T& value) {
-				char data[sizeof(T)];
-				for (int i = 0; i < sizeof(T); ++i) {
-					data[i] = reinterpret_cast<char*>(&value)[sizeof(T)-1-i];
-				}
-				return *reinterpret_cast<T*>(data);
-			}
-
-			enum class ENDENSE : uint16_t {
-				NONE = 0,
-				SAME = 0x0001,
-				DIFFER = 0x0100
-			} endien = ENDENSE::NONE;
-			uint32_t inkVersionNumber = 0;
-			uint32_t inkCppVersionNumber = 0;
-		};
 
 #ifdef INK_ENABLE_STL
 		story_impl(const char* filename);
