+++ inkcpp_compiler/emitter.h
@@ -2,6 +2,7 @@
 
 #include "command.h"
 #include "system.h"
+#include "version.h"
 #include "reporter.h"
 #include <string>
 #include <vector>
