+++ shared/public/version.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "system.h"
+
+namespace ink {
+	constexpr uint32_t VERSION = 0;
+};
